+++ app/cli/internal/action/attestation_push.go
@@ -25,10 +25,8 @@ import (
 	pb "github.com/chainloop-dev/chainloop/app/controlplane/api/controlplane/v1"
 	"github.com/chainloop-dev/chainloop/internal/attestation/crafter"
 	"github.com/chainloop-dev/chainloop/internal/attestation/renderer"
-	clsigner "github.com/chainloop-dev/chainloop/internal/attestation/signer"
+	"github.com/chainloop-dev/chainloop/internal/attestation/signer"
 	"github.com/secure-systems-lab/go-securesystemslib/dsse"
-	sigstoresigner "github.com/sigstore/sigstore/pkg/signature"
-	sigdsee "github.com/sigstore/sigstore/pkg/signature/dsse"
 	"google.golang.org/grpc"
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
@@ -134,13 +132,7 @@ func (action *AttestationPush) Run(ctx context.Context, attestationID string, ru
 	// Indicate that we are done with the attestation
 	action.c.CraftingState.Attestation.FinishedAt = timestamppb.New(time.Now())
 
-	var signer sigstoresigner.Signer
-	if action.keyPath != "" {
-		signer = clsigner.NewCosignSigner(action.keyPath, action.Logger)
-	} else {
-		signer = clsigner.NewChainloopSigner(action.keyPath, pb.NewSigningServiceClient(action.CPConnection), action.Logger)
-	}
-	wrappedSigner := sigdsee.WrapSigner(signer, "application/vnd.in-toto+json")
+	wrappedSigner := signer.GetSigner(action.keyPath, action.Logger, pb.NewSigningServiceClient(action.CPConnection))
 	renderer, err := renderer.NewAttestationRenderer(action.c.CraftingState, action.cliVersion, action.cliDigest, wrappedSigner,
 		renderer.WithLogger(action.Logger))
 	if err != nil {
