+++ internal/attestation/signer/chainloop/chainloop.go
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package signer
+package chainloop
 
 import (
 	"context"
